Api response modification and validations for all endpoints

+++ controllers/productController.js
@@ -1,4 +1,5 @@
 const Product = require("../models/productModel");
+const ResUtil = require("../utils/res");
 
 const getProducts = async (req, res) => {
   const { page = 1, limit = 10 } = req.query;
@@ -7,53 +8,97 @@ const getProducts = async (req, res) => {
       .limit(limit * 1)
       .skip((page - 1) * limit)
       .exec();
-    res.json(products);
+    ResUtil.SUCCESS(req, res, { products }, "SUCCESS");
   } catch (error) {
-    res.status(400).json({ error: error.message });
+    ResUtil.VALIDATION_ERROR(req, res, { error: error.message }, "ERROR");
   }
 };
 
 const getProduct = async (req, res) => {
   try {
+    if (!req.params.id) {
+      return ResUtil.VALIDATION_ERROR(
+        req,
+        res,
+        { error: "PRODUCT_ID_REQUIRED" },
+        "ERROR"
+      );
+    }
     const product = await Product.findById(req.params.id);
-    if (!product) return res.status(404).json({ error: "Product not found" });
-    res.json(product);
+    if (!product)
+      return ResUtil.VALIDATION_ERROR(
+        req,
+        res,
+        { error: "PRODUCT_NOT_FOUND" },
+        "ERROR"
+      );
+    ResUtil.SUCCESS(req, res, { product }, "SUCCESS");
   } catch (error) {
-    res.status(400).json({ error: error.message });
+    ResUtil.VALIDATION_ERROR(req, res, { error: error.message }, "ERROR");
   }
 };
 
 const createProduct = async (req, res) => {
+  const { name, description, price, stock } = req.body;
   try {
+    if (!name && !description && !price && !stock) {
+      return ResUtil.VALIDATION_ERROR(
+        req,
+        res,
+        { error: "ALL_FIELDS_ARE_REQUIRED" },
+        "ERROR"
+      );
+    }
     const product = await Product.create(req.body);
-    res.status(201).json(product);
+    ResUtil.SUCCESS(req, res, { product }, "SUCCESS");
   } catch (error) {
-    res.status(400).json({ error: error.message });
+    ResUtil.VALIDATION_ERROR(req, res, { error: error.message }, "ERROR");
   }
 };
 
 const updateProduct = async (req, res) => {
   try {
+    if (!req.params.id) {
+      return ResUtil.VALIDATION_ERROR(
+        req,
+        res,
+        { error: "PRODUCT_ID_REQUIRED" },
+        "ERROR"
+      );
+    }
     const product = await Product.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
     });
-    res.json(product);
+    ResUtil.SUCCESS(req, res, { product }, "SUCCESS");
   } catch (error) {
-    res.status(400).json({ error: error.message });
+    ResUtil.VALIDATION_ERROR(req, res, { error: error.message }, "ERROR");
   }
 };
 
 const deleteProduct = async (req, res) => {
   try {
+    if (!req.params.id) {
+      return ResUtil.VALIDATION_ERROR(
+        req,
+        res,
+        { error: "PRODUCT_ID_REQUIRED" },
+        "ERROR"
+      );
+    }
     const product = await Product.findByIdAndDelete(req.params.id);
-    
+
     if (!product) {
-      return res.status(404).send("Product not found");
+      return ResUtil.VALIDATION_ERROR(
+        req,
+        res,
+        { error: "PRODUCT_NOT_FOUND" },
+        "ERROR"
+      );
     }
 
-    res.status(200).send("Product deleted successfully.");
+    ResUtil.SUCCESS(req, res, {}, "PRODUCT_DELETED_SUCCESSFULLY");
   } catch (error) {
-    res.status(400).json({ error: error.message });
+    ResUtil.VALIDATION_ERROR(req, res, { error: error.message }, "ERROR");
   }
 };
